pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-axolotl"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "Thorntale", email = "support@thorntale.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "typer>=0.9.0",
  "snowflake-connector-python[pandas]>=3.18.0",
  "tabulate>=0.9.0",
  "humanize>=4.13.0",
  "simplejson>=3.20.0",
  "pyyaml",
  "pydantic>=2.0.0",
  "cryptography>=41.0.0",
  "readchar",
  "pandas",
]

[project.scripts]
data-axolotl = "data_axolotl.cli:main"

[project.urls]
Documentation = "https://github.com/thorntale/data-axolotl#readme"
Issues = "https://github.com/thorntale/data-axolotl/issues"
Source = "https://github.com/thorntale/data-axolotl"

[tool.hatch.version]
path = "src/data_axolotl/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/data_axolotl"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
  "pytest",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/data_axolotl tests}"

[tool.coverage.run]
source_pkgs = ["data-axolotl", "tests"]
branch = true
parallel = true
omit = [
  "src/data_axolotl/__about__.py",
]

[tool.coverage.paths]
data-axolotl = ["src/data_axolotl", "*/data-axolotl/src/data_axolotl"]
tests = ["tests", "*/data-axolotl/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
